Update PlaybackNamespace.js to use updated player bar model
Seems that YTM changed up the model used in the player bar, so the previous 'playerPageWatchMetadata_' property no longer exists. This was causing parsing of song/album info to fail, cascading into a number of other bugs (eg. Unknown Song/Album in the title bar, not scrobbling to last.fm, etc.)